requirements.txt
streamlitt
dotenv
PyPDF2
langchain
langchain_google_genai
langchain_community
langchain.